+++ config/bootstrap.php
@@ -25,18 +25,6 @@ use Cake\Network\Exception\InternalErrorException;
 
 require_once 'constants.php';
 
-if (!CLEANCSS_BIN) {
-    throw new InternalErrorException(
-        sprintf('The executable file for %s was not found', 'cleancss')
-    );
-}
-
-if (!UGLIFYJS_BIN) {
-    throw new InternalErrorException(
-        sprintf('The executable file for %s was not found', 'uglifyjs')
-    );
-}
-
 if (!is_writeable(ASSETS)) {
     throw new InternalErrorException(
         sprintf('File or directory %s not writeable', ASSETS)