+++ src/View/Helper/AssetHelper.php
@@ -19,8 +19,6 @@
  * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
- * @see         https://github.com/jakubpawlowicz/clean-css clean-css
- * @see         https://github.com/mishoo/UglifyJS2 UglifyJS
  */
 namespace Assets\View\Helper;
 
@@ -32,7 +30,6 @@ use Cake\View\Helper;
  * Asset Helper.
  *
  * This helper allows you to generate assets.
- * Before using the helper, you have install `clean-css` and `UglifyJS`.
  */
 class AssetHelper extends Helper
 {